+++ evennia/commands/default/admin.py
@@ -25,7 +25,7 @@ class CmdBoot(COMMAND_DEFAULT_CLASS):
     kick an account from the server.
 
     Usage
-      @boot[/switches] <account obj> [: reason]
+      boot[/switches] <account obj> [: reason]
 
     Switches:
       quiet - Silently boot without informing account
@@ -35,7 +35,7 @@ class CmdBoot(COMMAND_DEFAULT_CLASS):
     supplied it will be echoed to the user unless /quiet is set.
     """
 
-    key = "@boot"
+    key = "boot"
     switch_options = ("quiet", "sid")
     locks = "cmd:perm(boot) or perm(Admin)"
     help_category = "Admin"
@@ -46,7 +46,7 @@ class CmdBoot(COMMAND_DEFAULT_CLASS):
         args = self.args
 
         if not args:
-            caller.msg("Usage: @boot[/switches] <account> [:reason]")
+            caller.msg("Usage: boot[/switches] <account> [:reason]")
             return
 
         if ':' in args:
@@ -104,15 +104,19 @@ class CmdBoot(COMMAND_DEFAULT_CLASS):
 IPREGEX = re.compile(r"[0-9*]{1,3}\.[0-9*]{1,3}\.[0-9*]{1,3}\.[0-9*]{1,3}")
 
 
-def list_bans(banlist):
+def list_bans(cmd, banlist):
     """
     Helper function to display a list of active bans. Input argument
-    is the banlist read into the two commands @ban and @unban below.
+    is the banlist read into the two commands ban and unban below.
+
+    Args:
+        cmd (Command): Instance of the Ban command.
+        banlist (list): List of bans to list.
     """
     if not banlist:
         return "No active bans were found."
 
-    table = self.style_table("|wid", "|wname/ip", "|wdate", "|wreason")
+    table = cmd.styled_table("|wid", "|wname/ip", "|wdate", "|wreason")
     for inum, ban in enumerate(banlist):
         table.add_row(str(inum + 1),
                       ban[0] and ban[0] or ban[1],
@@ -125,7 +129,7 @@ class CmdBan(COMMAND_DEFAULT_CLASS):
     ban an account from the server
 
     Usage:
-      @ban [<name or ip> [: reason]]
+      ban [<name or ip> [: reason]]
 
     Without any arguments, shows numbered list of active bans.
 
@@ -133,7 +137,7 @@ class CmdBan(COMMAND_DEFAULT_CLASS):
     reason to be able to later remember why the ban was put in place.
 
     It is often preferable to ban an account from the server than to
-    delete an account with @accounts/delete. If banned by name, that account
+    delete an account with accounts/delete. If banned by name, that account
     account can no longer be logged into.
 
     IP (Internet Protocol) address banning allows blocking all access
@@ -141,10 +145,10 @@ class CmdBan(COMMAND_DEFAULT_CLASS):
     wildcard.
 
     Examples:
-      @ban thomas             - ban account 'thomas'
-      @ban/ip 134.233.2.111   - ban specific ip address
-      @ban/ip 134.233.2.*     - ban all in a subnet
-      @ban/ip 134.233.*.*     - even wider ban
+      ban thomas             - ban account 'thomas'
+      ban/ip 134.233.2.111   - ban specific ip address
+      ban/ip 134.233.2.*     - ban all in a subnet
+      ban/ip 134.233.*.*     - even wider ban
 
     A single IP filter can be easy to circumvent by changing computers
     or requesting a new IP address. Setting a wide IP block filter with
@@ -153,8 +157,8 @@ class CmdBan(COMMAND_DEFAULT_CLASS):
     or region.
 
     """
-    key = "@ban"
-    aliases = ["@bans"]
+    key = "ban"
+    aliases = ["bans"]
     locks = "cmd:perm(ban) or perm(Developer)"
     help_category = "Admin"
 
@@ -178,7 +182,7 @@ class CmdBan(COMMAND_DEFAULT_CLASS):
         if not self.args or (self.switches and
                              not any(switch in ('ip', 'name')
                                      for switch in self.switches)):
-            self.caller.msg(list_bans(banlist))
+            self.caller.msg(list_bans(self, banlist))
             return
 
         now = time.ctime()
@@ -214,15 +218,15 @@ class CmdUnban(COMMAND_DEFAULT_CLASS):
     remove a ban from an account
 
     Usage:
-      @unban <banid>
+      unban <banid>
 
-    This will clear an account name/ip ban previously set with the @ban
+    This will clear an account name/ip ban previously set with the ban
     command.  Use this command without an argument to view a numbered
     list of bans. Use the numbers in this list to select which one to
     unban.
 
     """
-    key = "@unban"
+    key = "unban"
     locks = "cmd:perm(unban) or perm(Developer)"
     help_category = "Admin"
 
@@ -232,7 +236,7 @@ class CmdUnban(COMMAND_DEFAULT_CLASS):
         banlist = ServerConfig.objects.conf('server_bans')
 
         if not self.args:
-            self.caller.msg(list_bans(banlist))
+            self.caller.msg(list_bans(self, banlist))
             return
 
         try:
@@ -261,9 +265,9 @@ class CmdEmit(COMMAND_DEFAULT_CLASS):
     admin command for emitting message to multiple objects
 
     Usage:
-      @emit[/switches] [<obj>, <obj>, ... =] <message>
-      @remit           [<obj>, <obj>, ... =] <message>
-      @pemit           [<obj>, <obj>, ... =] <message>
+      emit[/switches] [<obj>, <obj>, ... =] <message>
+      remit           [<obj>, <obj>, ... =] <message>
+      pemit           [<obj>, <obj>, ... =] <message>
 
     Switches:
       room     -  limit emits to rooms only (default)
@@ -272,12 +276,12 @@ class CmdEmit(COMMAND_DEFAULT_CLASS):
 
     Emits a message to the selected objects or to
     your immediate surroundings. If the object is a room,
-    send to its contents. @remit and @pemit are just
-    limited forms of @emit, for sending to rooms and
+    send to its contents. remit and pemit are just
+    limited forms of emit, for sending to rooms and
     to accounts respectively.
     """
-    key = "@emit"
-    aliases = ["@pemit", "@remit"]
+    key = "emit"
+    aliases = ["pemit", "remit"]
     switch_options = ("room", "accounts", "contents")
     locks = "cmd:perm(emit) or perm(Builder)"
     help_category = "Admin"
@@ -290,9 +294,9 @@ class CmdEmit(COMMAND_DEFAULT_CLASS):
 
         if not args:
             string = "Usage: "
-            string += "\n@emit[/switches] [<obj>, <obj>, ... =] <message>"
-            string += "\n@remit           [<obj>, <obj>, ... =] <message>"
-            string += "\n@pemit           [<obj>, <obj>, ... =] <message>"
+            string += "\nemit[/switches] [<obj>, <obj>, ... =] <message>"
+            string += "\nremit           [<obj>, <obj>, ... =] <message>"
+            string += "\npemit           [<obj>, <obj>, ... =] <message>"
             caller.msg(string)
             return
 
@@ -301,10 +305,10 @@ class CmdEmit(COMMAND_DEFAULT_CLASS):
         send_to_contents = 'contents' in self.switches
 
         # we check which command was used to force the switches
-        if self.cmdstring == '@remit':
+        if self.cmdstring == 'remit':
             rooms_only = True
             send_to_contents = True
-        elif self.cmdstring == '@pemit':
+        elif self.cmdstring == 'pemit':
             accounts_only = True
 
         if not self.rhs:
@@ -341,12 +345,12 @@ class CmdNewPassword(COMMAND_DEFAULT_CLASS):
     change the password of an account
 
     Usage:
-      @userpassword <user obj> = <new password>
+      userpassword <user obj> = <new password>
 
     Set an account's password.
     """
 
-    key = "@userpassword"
+    key = "userpassword"
     locks = "cmd:perm(newpassword) or perm(Admin)"
     help_category = "Admin"
 
@@ -356,7 +360,7 @@ class CmdNewPassword(COMMAND_DEFAULT_CLASS):
         caller = self.caller
 
         if not self.rhs:
-            self.msg("Usage: @userpassword <user obj> = <new password>")
+            self.msg("Usage: userpassword <user obj> = <new password>")
             return
 
         # the account search also matches 'me' etc.
@@ -388,8 +392,8 @@ class CmdPerm(COMMAND_DEFAULT_CLASS):
     set the permissions of an account/object
 
     Usage:
-      @perm[/switch] <object> [= <permission>[,<permission>,...]]
-      @perm[/switch] *<account> [= <permission>[,<permission>,...]]
+      perm[/switch] <object> [= <permission>[,<permission>,...]]
+      perm[/switch] *<account> [= <permission>[,<permission>,...]]
 
     Switches:
       del     -  delete the given permission from <object> or <account>.
@@ -398,8 +402,8 @@ class CmdPerm(COMMAND_DEFAULT_CLASS):
     This command sets/clears individual permission strings on an object
     or account. If no permission is given, list all permissions on <object>.
     """
-    key = "@perm"
-    aliases = "@setperm"
+    key = "perm"
+    aliases = "setperm"
     switch_options = ("del", "account")
     locks = "cmd:perm(perm) or perm(Developer)"
     help_category = "Admin"
@@ -412,7 +416,7 @@ class CmdPerm(COMMAND_DEFAULT_CLASS):
         lhs, rhs = self.lhs, self.rhs
 
         if not self.args:
-            string = "Usage: @perm[/switch] object [ = permission, permission, ...]"
+            string = "Usage: perm[/switch] object [ = permission, permission, ...]"
             caller.msg(string)
             return
 
@@ -496,19 +500,19 @@ class CmdWall(COMMAND_DEFAULT_CLASS):
     make an announcement to all
 
     Usage:
-      @wall <message>
+      wall <message>
 
     Announces a message to all connected sessions
     including all currently unlogged in.
     """
-    key = "@wall"
+    key = "wall"
     locks = "cmd:perm(wall) or perm(Admin)"
     help_category = "Admin"
 
     def func(self):
         """Implements command"""
         if not self.args:
-            self.caller.msg("Usage: @wall <message>")
+            self.caller.msg("Usage: wall <message>")
             return
         message = "%s shouts \"%s\"" % (self.caller.name, self.args)
         self.msg("Announcing to all connected sessions ...")
@@ -520,12 +524,12 @@ class CmdForce(COMMAND_DEFAULT_CLASS):
     forces an object to execute a command
 
     Usage:
-        @force <object>=<command string>
+        force <object>=<command string>
 
     Example:
-        @force bob=get stick
+        force bob=get stick
     """
-    key = "@force"
+    key = "force"
     locks = "cmd:perm(spawn) or perm(Builder)"
     help_category = "Building"
     perm_used = "edit"